py/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
chromadb>=0.4.0
dashscope>=1.0.0
